+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.0.38
+
+* Add support of `Toggle`
+* Attributes collection now can be concatenated
+
 ## 0.0.37
 
 * `Dialog` now do not add auto margin by default